+++ modules/database.py
@@ -1,0 +1,19 @@
+import sqlite3
+
+DB_NAME = 'habitmaster.db'
+
+def create_database():
+    conn = sqlite3.connect(DB_NAME)
+    c = conn.cursor()
+
+    # Create habit table
+    c.execute('''
+        CREATE TABLE IF NOT EXISTS habits (
+            id INTEGER PRIMARY KEY AUTOINCREMENT,
+            name TEXT NOT NULL,
+            completed INTEGER DEFAULT 0
+        )
+    ''')
+
+    conn.commit()
+    conn.close()
